More graphics updates, and make install fix for installing on top of an old install (templates were incorrectly copied if system-wide blitwizard templates folder was already present)

+++ src/graphicssdltexturestruct.h
@@ -26,7 +26,7 @@
 
 struct graphicstexture {
     // basic info
-    unsigned int width,height;
+    size_t width,height;
     // SDL info
     SDL_Texture* sdltex;
 };
